+++ Makefile
@@ -91,6 +91,11 @@ ARROW_READ_CPP += $(ARROW_READ_FILE_NAME).cpp
 ARROW_READ_H += $(ARROW_READ_FILE_NAME).h
 ARROW_READ_O += $(ARKOUDA_SOURCE_DIR)/ReadParquet.o
 
+ARROW_WRITE_FILE_NAME += $(ARKOUDA_SOURCE_DIR)/parquet/WriteParquet
+ARROW_WRITE_CPP += $(ARROW_WRITE_FILE_NAME).cpp
+ARROW_WRITE_H += $(ARROW_WRITE_FILE_NAME).h
+ARROW_WRITE_O += $(ARKOUDA_SOURCE_DIR)/WriteParquet.o
+
 ARROW_OTHER_FILE_NAME += $(ARKOUDA_SOURCE_DIR)/parquet/OtherParquet
 ARROW_OTHER_CPP += $(ARROW_OTHER_FILE_NAME).cpp
 ARROW_OTHER_H += $(ARROW_OTHER_FILE_NAME).h
@@ -207,8 +212,9 @@ endif
 
 .PHONY: compile-arrow-cpp
 compile-arrow-cpp:
-	$(CHPL_CXX) -O3 -std=c++17 -c $(ARROW_READ_CPP) -o $(ARROW_READ_O) $(INCLUDE_FLAGS) $(ARROW_SANITIZE)
 	$(CHPL_CXX) -O3 -std=c++17 -c $(ARROW_OTHER_CPP) -o $(ARROW_OTHER_O) $(INCLUDE_FLAGS) $(ARROW_SANITIZE)
+	$(CHPL_CXX) -O3 -std=c++17 -c $(ARROW_READ_CPP) -o $(ARROW_READ_O) $(INCLUDE_FLAGS) $(ARROW_SANITIZE)
+	$(CHPL_CXX) -O3 -std=c++17 -c $(ARROW_WRITE_CPP) -o $(ARROW_READ_O) $(INCLUDE_FLAGS) $(ARROW_SANITIZE)
 
 $(ARROW_O): $(ARROW_CPP) $(ARROW_H)
 	make compile-arrow-cpp
